Renamed Scheduler to Task, use duration instead of int millis
