+++ src/main/scala/nl/jqno/secretaresse/Task.scala
@@ -1,0 +1,28 @@
+package nl.jqno.secretaresse
+
+import java.util.{Timer, TimerTask}
+
+import scala.concurrent.duration._
+
+class Task(supplier: Unit => Unit) {
+
+  private var timer = new Timer()
+
+  def runEvery(period: Duration) = {
+    stop()
+    timer.schedule(new TimerTask {
+      override def run(): Unit = supplier()
+    }, (1 second).toMillis, period.toMillis)
+
+  }
+
+  def stop(): Unit = {
+    timer.cancel()
+    timer.purge()
+    timer = new Timer()
+  }
+}
+
+object Task {
+  def apply(task: Unit => Unit): Task = new Task(task)
+}
